[Chore #147018467]Change file structure.
