+++ my_app/__init__.py
@@ -1,13 +1,11 @@
 import os
-from flask import request, jsonify, abort
-from flask_api import FlaskAPI
-from flask import make_response
-from flask_sqlalchemy import SQLAlchemy
-from instance.config import app_config
-from my_app.models import User, Bucketlist, BucketlistItem
-from my_app.models import db
-from .auth import auth_blueprint
 
+from flask_api import FlaskAPI
+from flask_sqlalchemy import SQLAlchemy
+
+from instance.config import app_config
+from my_app.models import db
+from my_app.views import auth_blueprint, bucketlist_blueprint, bucketlist_item_blueprint
 
 def create_app(config_name):
     app = FlaskAPI(__name__, instance_relative_config=True)
@@ -17,71 +15,8 @@ def create_app(config_name):
     app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://localhost/Bucketlistdb'
     db.init_app(app)
 
-    @app.route('/bucketlists/', methods=['POST', 'GET'])
-    def bucketlists():
-        if request.method == "POST":
-            name = str(request.data.get('name', ''))
-            if name:
-                bucketlist = Bucketlist(name=name)
-                bucketlist.save()
-                response = jsonify({
-                    'id': bucketlist.id,
-                    'name': bucketlist.name,
-                    'created_by': bucketlist.created_by,
-                    'date_created': bucketlist.date_created,
-                    'date_modified': bucketlist.date_modified
-                })
-                response.status_code = 201
-                return response
-        else:
-            bucketlists = Bucketlist.get_all()
-            results = []
-
-            for bucketlist in bucketlists:
-                obj = {
-                    'id': bucketlist.id,
-                    'name': bucketlist.name,
-                    'date_created': bucketlist.date_created,
-                    'date_modified': bucketlist.date_modified
-                }
-                results.append(obj)
-            response = jsonify(results)
-            response.status_code = 200
-            return response
-
-    @app.route('/bucketlists/<int:id>', methods=['GET', 'PUT', 'DELETE'])
-    def bucketlist_manipulation(id, **kwargs):
-        bucketlist = Bucketlist.query.filter_by(id=id).first()
-        if not bucketlist:
-            abort(404)
-            return make_response(jsonify(response)), 404
-
-        if request.method == 'DELETE':
-            bucketlist.delete()
-            return {
-            "message": "bucketlist {} deleted successfully".format(bucketlist.id)
-         }, 200
-
-        elif request.method == 'PUT':
-            name = str(request.data.get('name', ''))
-            bucketlist.name = name
-            bucketlist.save()
-            response = jsonify({
-                'id': bucketlist.id,
-                'name': bucketlist.name,
-                'date_created': bucketlist.date_created,
-                'date_modified': bucketlist.date_modified
-            })
-            response.status_code = 200
-            return response
-        else:
-            response = jsonify({
-                'id': bucketlist.id,
-                'name': bucketlist.name,
-                'date_created': bucketlist.date_created,
-                'date_modified': bucketlist.date_modified
-            })
-            response.status_code = 200
-            return response
     app.register_blueprint(auth_blueprint)
+    app.register_blueprint(bucketlist_blueprint)
+    app.register_blueprint(bucketlist_item_blueprint)
+
     return app
