+++ tests/test_user_authentication.py
@@ -2,7 +2,6 @@ import unittest
 import json
 from my_app import create_app, db
 
-
 class AuthTestCase(unittest.TestCase):
     """Test case for the authentication blueprint."""
